fix: フォントをビルドするかどうかを事前にチェックするように

+++ bin/create_font_subset.sh
@@ -2,10 +2,28 @@
 
 PWD=$(pwd)
 
+function uniq_chars_from_file() {
+  echo `cat $1 | sed "s/\(.\)/\1\n/g" | sort | LC_COLLATE=C uniq | tr -d "\n"`
+}
+
+recent=$(uniq_chars_from_file .cache/chars.txt)
+now=$(uniq_chars_from_file out/index.html)
+
+if [ -f ".cache/chars.txt" -a "$recent" = "$now" ]; then
+  echo "font file is newest, font building was skipped."
+  exit 0
+fi
+
+echo "font file is older, font bulding..."
+
 mkdir -p .temp
 rm -rf .temp/*
 
 docker build -t happou31/font-create .
-cat out/index.html | sed "s/\(.\)/\1\n/g" | sort | LC_COLLATE=C uniq | tr -d "\n" > .temp/chars.txt
+echo $now > .temp/chars.txt
 docker run --rm -v $PWD/.temp/:/root/ happou31/font-create
-mv .temp/sub_nicomoji-plus.woff2 static/font/
+
+mv -f .temp/sub_nicomoji-plus.woff2 static/font/
+
+mkdir -p .cache
+echo $now > .cache/chars.txt
